+++ core-tests/Timeouts.hs
@@ -4,9 +4,6 @@ import Test.Tasty
 import Test.Tasty.Runners
 import Test.Tasty.HUnit
 import Control.Concurrent
-#if !MIN_VERSION_base(4,8,0)
-import Data.Monoid (mempty)
-#endif
 import Utils
 
 -- this is a dummy tree we use for testing