+++ core/Test/Tasty/Options.hs
@@ -33,7 +33,7 @@ import Data.Monoid
 import Data.Foldable
 import Prelude hiding (mod) -- Silence FTP import warnings
 import Options.Applicative
-#if MIN_VERSION_base(4,9,0)
+#if !MIN_VERSION_base(4,11,0)
 import Data.Semigroup (Semigroup)
 import qualified Data.Semigroup (Semigroup((<>)))
 #endif
@@ -88,10 +88,8 @@ instance Monoid OptionSet where
   mempty = OptionSet mempty
   OptionSet a `mappend` OptionSet b =
     OptionSet $ Map.unionWith (flip const) a b
-#if MIN_VERSION_base(4,9,0)
 instance Semigroup OptionSet where
   (<>) = mappend
-#endif
 
 -- | Set the option value
 setOption :: IsOption v => v -> OptionSet -> OptionSet