+++ core/Test/Tasty/Patterns/Eval.hs
@@ -10,10 +10,6 @@ import Data.List
 import Data.Maybe
 import Data.Char
 import Test.Tasty.Patterns.Types
-#if !MIN_VERSION_base(4,8,0)
-import Control.Applicative
-import Data.Traversable
-#endif
 
 type Path = Seq.Seq String